+++ CHANGELOG.md
@@ -37,6 +37,12 @@
 - Fix incorrect error handling during cluster CA secret creation
   [#2250](https://github.com/Kong/kong-operator/pull/2250)
 
+### Changed
+
+- For Hybrid `Gateway`s the operator does not run the `ControlPlane` anymore, as
+  the `DataPlane` is configured to use `Koko` as Konnect control plane.
+  [#2253](https://github.com/Kong/kong-operator/pull/2253)
+
 ## [v2.0.0]
 
 > Release date: 2025-09-09